Move period and assignment operators to control characters and make all other operators as lists.
We do this because irrespective of which operator it is we still just consider
it as only operator in the parse tree construction. It is only while
translating it to intermediate representation, we will have to worry which
operator exactly it is and for that we need not parse it using separate parser
method for each operator, they can all be parsed with one single generic operator
method.