Reimplement the whole LiveIntervalsHeap.
There was a major mistake in my code which was storing a single key value pair
per (instruction, register). However this has two problems. First being an
instruction can hold only one register. In order to solve this problem, the
(key, value) pair was changed to (instruction, list of registers). So now
the second problem comes into play. If we build our heap list with just this
dictionary.keys() since every key appears only once we get a single entry in
the heap even for instructions which have multiple registers in the list.
However since we have to check liveness range for each register and sorted
by the order of these start instruction keys in the dictionary we will want
an entry in the heap for every register.

So this new approach will rebuild this heap structure by repeating the number
of values the key has (i.e. the number of registers that start at an instruction)
every time we construct the heap. This has been changed in all the methods
abstracting the heap.

In addition to this the whole datastructure implementation is completed with
iterator protocol and track of the last element popped.