Keep track of the last register count in the function.
This is a bit contrived name because the count is actually the count.
The name of the last register is -1 of this value. So while assigning
the name for this register we can directly use this count first and
then increment the counter.

This needs to be fed to the start node of the function. But we know this
count only when we hit the last instruction in the function. So we initially
set this count when we know it at the .end_<function_name> instruction
within that instruction object temporarily. Finally when we analyze basic
blocks, we set this value in the start node of the basic block for easy
retrieval.

The start node of basic blocks for a function has more or less evolved into
a node that stores metadata about function. This is mostly because of my
stupidity for not compiling each functions independently.