Use the last register count variable to assign names to the new registers that are caused by spilling.
This is to ensure that the names don't overlap. Although this should not matter
since they are different objects, it makes it painful for debugging when we
see the same register names for two registers if the counter restarts somewhere
in the middle of the function when they are clearly not even related. Note these
two common name registers are not necessarily the spilled register input and
results. Since spilling register r4 first will create a new r0 spilled register
if we reset the register counters.