There is no reason to use a string to represent a register's bit variable, so use a tuple so that accesses are easier.
Note, by doing this we can have any innumerable number of registers in the real
machine register space. Earlier when we used strings, we used to divide the string
to, from first to last but one character as the register bit variable name and
the last character as the bit position. This restricted us to 1 character to
represent bit position which meant with 10 values, we could only represent 2 ^ 10,
i.e. 1024 registers. So if some real machine allows more than 1024 registers sometime
in the future, this would not work. But now it should!