Total misconception about phi instructions, fixed major conceptual mistake.
All the while I was assuming that the phi operands must be alive
after the instruction is executed, while I did not do this for
any other instructions. An operand which is the last use never
collides with the result of that instruction since the result of
the instruction can be assigned the same register as this operand
that goes dead.

yAy! This bug has been fixed now.