Keep track of phi operands so as to resolve for loop invariants.
Loop invariants are live-in at loop header and if used inside the loop
and spilled, they must reloaded so we need to make sure to insert move
or reload instructions.

NOTE: We introduce an instruction called mmove instruction which is not
part of the IR that Prof. Franz has given, but I don't know of any better
way of keeping track of memory to memory moves. This happens when the
loop invariant is spilled twice, once before entering the loop and again
after the last use in the loop.