Now I see this file and compare it with pydevtools and this file is so different so removing the related lines since they are not even similar.
This structure is not anybody else idea. So doesn't make sense to put that in
the credit since that would be wrong. Not that I don't want to thank the authors
of pydevtools. I still heartily thank them for helping get the grip of ELF format.
But this file is completely different from what pydevtools has.