Implement the loop detection algorithm using the dominator tree.
It was stupid of me to not think about this before. What we now do is
if there is an edge from a node to one of the already visited node,
then this already visited node is a candidate loop header and the
current node in question is a candidate loop footer, check if the
candidate header dominates the candidate footer (or in other words),
starting at the subtree of the dominator tree rooted at the candidate
loop header node, see if the candidate loop footer is present in the
subtree if so they are indeed the loop header-footer pair and hence
mark them.

I came to know after implementing this algorithm that Dragon book
gives exactly the same algorithm! Not bad!