Hand written binary strings for InputNum, OutputNum and OutputNewline functions.
These binary strings were completely handwritten without any additional support
except the Intel x86_64 spec sheet and looking at the GCC generated binary disassembly
for the addressing mode used for indirect addressing of the %rsp register.

And why did I hand write this binary? Because I have gone mad. Totally gone mad! :D

On a more serious note, that was not an emotional imbalance ;-) It gives me this
immense pleasure, which I cannot put down in words, to do anything computer science.
It is my dream to be able to brag that I can do anything computer science/programming!
Be it writing web applications (which is seen as something easy or cheap today while it
is not), or be it writing Desktop applications or write a more system level stuff like
Filesystems, drivers etc. or writing real machine binaries!

Also, I love when I can be in total control of the machine! "The Master who is in control
of the machine" and I totally love it when I can issue the instructions to my machine.
This code is also a result of that. Computer is after all a machine!

In addition to all these I have been having this dream from quite a long time to
bootstrap my own compiler from step 1, i.e. handwrite binaries to build an assembler
and use it to build a compiler that compiles my own language. This is one step closer
to that dream! I hope I can take this further!