Fix the way DIV instructions code is generated.
DIV instruction in x86 is a big mess. This is the worsht thing ever
that DIV instructions don't take immediate values as operands and
the dividend should be in %rdx:%rax. I have no clue why one would
design an instruction like that. But anyway, code generator is
rewritten for DIV instruction to take care of that.

Further Division by an immediate is implemented using, what I call
a weird algorithm which computes the approximation of 1 / x first
and then multiplies with the dividend in both GCC and LLVM. I really
don't have patience to implement that algorithm. Intel, come'on guys,
grow up!