Fix arithematic instructions so that, if the result and one of the operands are same, they don't get overwritten.
With the strategy we were using before this commit, if the second operand of
the instruction had the same register assigned as the result, we used to
arbitrarily choose first argument to move to result to convert Three Address
Code to x86's Two Address Code without looking at the second operand and if
the second operand was actually the result operand also, this would overwrite
the second operand even before the operation. This is fixed now.