Fix the %rax register pushing and popping at call handling to avoid %rax overwriting.
If the result of the call instruction is to be stored in %rax we need not
push or pop it from the stack while doing a function call. In other cases
we will have to push %rax first onto the stack and pop it in the end even
after the result is moved to its appropriate location to avoid overwriting.