When the offset of memory is immediate, i.e. for array accesses we still need to create a new memory object.
This is because manipulating the offset in the given memory object which is
shared by all other instructions which accesses the same base memory object
will overwrite the offsets for all the instructions which is not correct.

So in order to compute the offsets correctly for each instruction we will
have to create a new memory object for each instruction accessing such
memory.