Rewrite the Live Range Analyzer to use set unions instead of range merging.
Live Range Analysis was done using the ranges of instructions where
registers were alive. This had an inherent problem. When two ranges
were merged, it was merged using the smaller of the two ranges as
the start of the new range and the larger of the two ranges as the
end of the new range. This caused problems in cases of registers
whose live ranges ended in if-else blocks since we would take the
last most instruction label irrespective of where the live range
of the register ends. So unions are better approaches.