Teach Inteference Graph nodes about preferred edges.
All the phi-operand and phi-result pairs should be attempted to be assigned
the same color as far as possible. This is equivalent to coalescing. So in
order to achieve that Interference Graph should record that information which
is recorded via preferred edges.

Nodes which have the preferred edges are attempted to be assigned the same
color.