Build the clause strings to be supplied to MaxSAT solver in the reduction method.
Build the strings in the reduction method rather than deferring upto
the point when we have to write it to the DIMACS file. We have to do
this because it is here that we know the weights of individual clauses
exactly and it is easier to build the clause strings here along with
the weights instead of waiting until writing to the DIMACS file.

So we also count the number of clauses here and call the method that
builds the clauses for coalescing as well.